+++ remix/server/routes/api/[...].ts
@@ -165,6 +165,8 @@ export const routeModules: Record<string, () => Promise<RouteModule>> = {
   'v1/lopu/chats/update': () => import('../../../app/routes/api/v1/lopu/chats/update/_update'),
   'v1/lopu/chats/delete': () => import('../../../app/routes/api/v1/lopu/chats/delete/_delete'),
   'v1/lopu/chats/reply': () => import('../../../app/routes/api/v1/lopu/chats/reply/_reply'),
+	'v1/lopu/vault': () => import('../../../app/routes/api/v1/lopu/vault/_vault'),
+	'v1/lopu/voice/reply': () => import('../../../app/routes/api/v1/lopu/voice/reply/_reply'),
   'v1/mongodb/endpoint': () => import('../../../app/routes/api/v1/mongodb/endpoint/_endpoint'),
   'v1/mongodb/endpoints': () => import('../../../app/routes/api/v1/mongodb/endpoints/_endpoints'),
   'v1/mongodb/get-connection': () => import('../../../app/routes/api/v1/mongodb/get-connection/_get-connection'),